presunutie funkcii histogramu do separatneho suboru kvoli zvyseniu prehladnosti a velkosti funkcie
